daily_menu/requirements.txt
Django==2.1.4
pytz
pyzomato
requests
beautifulsoup4
django-debug-toolbar
dateparser
djangorestframework==3.9.2
psycopg2-binary
lxml
elasticsearch-dsl>=6.0.0,<7.0.0
django-cors-headers==2.5.2